1.暂时屏蔽自动登录； 2.push tag.

+++ Android/trunk/src/com/etoc/weflow/view/MagicTextView.java
@@ -22,13 +22,14 @@ public class MagicTextView extends TextView {
 
 	@TargetApi(11)
 	public void showNumberWithAnimation(String number, int duration) {
-		int num = 0;
+		float num = 0;
 		try {
-			num = Integer.parseInt(number);
+			num = Float.parseFloat(number);
 		} catch(Exception e) {
+			
 			e.printStackTrace();
 		}
-		showNumberWithAnimation(num, duration);
+		showNumberWithAnimation((int)num, duration);
 	}
 	
 	@TargetApi(11)
